ZHA: Refactor with KBS
