+++ with-docker-compose-app/next-app/src/components/Footer/Footer.module.css
@@ -1,8 +1,6 @@
 .footer {
     border-top-right-radius: 4rem; 
     background-color: var( --very-dark-black-blue);
-    margin-top: 5px;
-    padding: 10px 5px;
 }
 
 .footer h6 {
