+++ with-docker-compose-app/next-app/src/components/Header/Header.module.css
@@ -16,7 +16,6 @@
 }
 
 .navBar {
-    padding: 3rem 8rem;
     gap: 6rem;
 }
