+++ with-docker-compose-app/next-app/src/components/Header/Logo.js
@@ -10,7 +10,7 @@ function Logo() {
             src="/images/logo.svg"
             alt="Blogr"
             width={100}
-            height={50}
+            height={40}
             className={styles.logo}
           />
     </Link>
