--- with-docker-compose-app/next-app/src/components/layout/layout.js
@@ -1,15 +1,0 @@
-import React from 'react'
-import Footer from './footer'; 
-import Header from './Header/Header'; 
-
-function Layout() {
-    return (
-        <>
-            <Header />
-            <main>Main section</main>
-            <Footer />
-        </>
-    )
-}
-
-export default Layout;
