+++ with-docker-compose-app/next-app/src/pages/layout.js
@@ -1,0 +1,12 @@
+import Header from "../components/Header/Header";
+import Footer from "../components/Footer/Footer";
+
+export default function Layout({ children }) {
+  return (
+    <>
+        <Header />
+            {children}
+        <Footer />
+    </>
+  )
+}
